requirements.txt
black
pytest
pytest-cov
pylint
python-datamuse
requests
streamlit
tensorflow
torch
transformers

